+++ aspnet-core/src/Gsv.Core/Objects/Shelf.cs
@@ -33,7 +33,7 @@ namespace Gsv.Objects
 
         public virtual CargoType CargoType { get; set; }
 
-        public float? Inventory { get; set; }
+        public double? Inventory { get; set; }
         public int NumInToday { get; set; }
         public int NumOutToday { get; set; }
         public DateTime? LastInTime { get; set; }
