+++ aspnet-core/src/Gsv.Core/Tasks/InStock.cs
@@ -33,7 +33,7 @@ namespace Gsv.Tasks
         public int ShelfId { get; set; }
         public virtual Shelf Shelf { get; set; }
 
-        public float Quantity { get; set; }
+        public double Quantity { get; set; }
 
         public int SourceId { get; set; }
         public virtual Source Source { get; set; }
